// src/types.ts
export type FlatTreeItem = {
  mapId: NodeKey;
  path: NumberOrStringArray;
  parentKey: NodeKey | null;
  draggingNode?: boolean;
  dropSuccessNode?: boolean;
  dropErrorNode?: boolean;
  forcedDepth?: number;
};

// Common Types
export type NumberOrStringArray = Array<string | number>;
export type ContextProviderProps = { children: React.JSX.Element };
export type NodeKey = string | number;
export type TreeMap = Record<NodeKey, TreeItem>;
export type SearchedNodeMap = Record<NodeKey, boolean>;

export type TreeItem<T = {}> = T & {
  id?: NodeKey;
  title?: React.ReactNode | undefined;
  subtitle?: React.ReactNode | undefined;
  expanded?: boolean | undefined;
  children?: Array<TreeItem<T>> | undefined;
};

export interface TreeNode<T = {}> {
  node: TreeItem<T>;
}

export interface TreePath {
  path: NumberOrStringArray;
}

export interface TreeIndex {
  treeIndex: number;
}
// ...
export type GetNodeKeyFn<T = {}> = (data: TreeNode<T> & TreeIndex) => NodeKey;